+++ src/pytest_locker/__init__.py
@@ -1,3 +1,3 @@
 __version__ = "0.1.0"
 
-from pytest_locker.fixtures import JsonLocker, Locker, json_locker, locker  # noqa
+from pytest_locker.fixtures import Locker, locker  # noqa
