+++ test/test_locker.py
@@ -4,7 +4,7 @@ from unittest.mock import mock_open, patch
 from _pytest.capture import CaptureFixture
 from pytest import raises
 
-from pytest_locker import JsonLocker, Locker
+from pytest_locker import Locker
 from pytest_locker.fixtures import UserDidNotAcceptDataException
 
 
@@ -75,35 +75,35 @@ def test_multiline_diff(locker: Locker, capsys: CaptureFixture) -> None:
         locker.lock(captured_out, "rhyme_diff")
 
 
-def test_json_locker_default(json_locker: JsonLocker) -> None:
-    json_locker.lock(
+def test_json_locker_default(locker: Locker) -> None:
+    locker.lock(
         {"description": "this is some text json", "items": ["items", "should", "work"]}
     )
 
 
-def test_json_locker_dataclass(json_locker: JsonLocker) -> None:
+def test_json_locker_dataclass(locker: Locker) -> None:
     from dataclasses import dataclass
 
     @dataclass
     class TestClass:
         description: str
 
-    json_locker.lock(TestClass(description="this is a test dataclass"))
+    locker.lock(TestClass(description="this is a test dataclass"))
 
 
-def test_json_locker_pydantic(json_locker: JsonLocker) -> None:
+def test_json_locker_pydantic(locker: Locker) -> None:
     from pydantic import BaseModel
 
     class TestClass(BaseModel):
         description: str
 
-    json_locker.lock(TestClass(description="this is a test dataclass"))
+    locker.lock(TestClass(description="this is a test dataclass"))
 
 
-def test_json_locker_unserializable_class(json_locker: JsonLocker) -> None:
+def test_json_locker_unserializable_class(locker: Locker) -> None:
     class TestClass:
         def __init__(self, description: str) -> None:
             self.description = description
 
     with raises(TypeError):
-        json_locker.lock(TestClass(description="this is a test dataclass"))
+        locker.lock(TestClass(description="this is a test dataclass"))
